requirements/development.txt
-r base.txt
blessings==1.6
bpython==0.14.2
coverage==3.7.1
curtsies==0.1.19
django-extensions==1.5.5
django-mongoengine==0.1
django-nose==1.4.1
lxml==3.4.4
nose==1.3.7
nose-mongoengine==0.2.2
Pygments==2.0.2
requests==2.7.0
